backend/requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy
stix2
taxii2-client
PyYAML
requests
requests-mock
